requirements.txt
# Benchmark / Games
pyyaml==6.0 # Logging
numpy==1.24.3 # Score computation
nltk==3.8.1 # Taboo
spacy==3.7.4 # Taboo
Levenshtein==0.25.0 # reference
tiktoken==0.4.0 # Wordle
# Evaluation
scikit-learn==1.2.2
matplotlib==3.7.1
pandas==2.0.1
seaborn==0.12.2
jupyter==1.0.0
# Backends
retry==0.9.2 # API call utility
aleph-alpha-client==7.0.1
openai==1.12.0
anthropic==0.16.0
cohere==4.48
google-generativeai==0.5.3
mistralai==0.0.12
# Slurk Integration
python-engineio==4.4.0
python-socketio==5.7.2
websocket-client
# multimodal games
imageio==2.34.1
networkx==3.2.1
# Models
torch==2.5.0
transformers==4.47.0
accelerate>=0.26.0
